# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MFT API",
    "version": "0.1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Health Health Get"
                }
              }
            }
          }
        }
      }
    },
    "/stats": {
      "get": {
        "tags": [
          "stats"
        ],
        "summary": "Get Stats",
        "operationId": "get_stats_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/apis": {
      "get": {
        "tags": [
          "apis"
        ],
        "summary": "List Apis",
        "operationId": "list_apis_apis_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/sts/strategies": {
      "get": {
        "tags": [
          "sts"
        ],
        "summary": "List Strategies",
        "operationId": "list_strategies_sts_strategies_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategiesResponse"
                }
              }
            }
          }
        }
      }
    },
    "/sts/sessions": {
      "get": {
        "tags": [
          "sts"
        ],
        "summary": "List Sessions",
        "operationId": "list_sessions_sts_sessions_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "live",
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sts/sessions/{session_id}/pause": {
      "post": {
        "tags": [
          "sts"
        ],
        "summary": "Pause Session",
        "operationId": "pause_session_sts_sessions__session_id__pause_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StsControlResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sts/sessions/{session_id}/resume": {
      "post": {
        "tags": [
          "sts"
        ],
        "summary": "Resume Session",
        "operationId": "resume_session_sts_sessions__session_id__resume_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StsControlResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sts/sessions/{session_id}/stop": {
      "post": {
        "tags": [
          "sts"
        ],
        "summary": "Stop Session",
        "operationId": "stop_session_sts_sessions__session_id__stop_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StsControlResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sts/{strategy_id}": {
      "post": {
        "tags": [
          "sts"
        ],
        "summary": "Deploy",
        "description": "Deploy a strategy; API orchestrates STS create + TD attach.",
        "operationId": "deploy_sts__strategy_id__post",
        "parameters": [
          {
            "name": "strategy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Strategy Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeployBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeployResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/td/sessions": {
      "get": {
        "tags": [
          "td"
        ],
        "summary": "List Sessions",
        "operationId": "list_sessions_td_sessions_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "live",
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/md/sessions": {
      "get": {
        "tags": [
          "md"
        ],
        "summary": "List Sessions",
        "operationId": "list_sessions_md_sessions_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "live",
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/audits": {
      "get": {
        "tags": [
          "audits"
        ],
        "summary": "List Audits",
        "operationId": "list_audits_audits_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ApiListResponse": {
        "properties": {
          "apis": {
            "items": {
              "$ref": "#/components/schemas/ApiOut"
            },
            "type": "array",
            "title": "Apis"
          }
        },
        "type": "object",
        "title": "ApiListResponse"
      },
      "ApiOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "venue": {
            "type": "string",
            "title": "Venue"
          },
          "api_key": {
            "type": "string",
            "title": "Api Key"
          },
          "type": {
            "type": "string",
            "title": "Type"
          }
        },
        "type": "object",
        "required": [
          "id",
          "venue",
          "api_key",
          "type"
        ],
        "title": "ApiOut"
      },
      "AuditListResponse": {
        "properties": {
          "audits": {
            "items": {
              "$ref": "#/components/schemas/AuditOut"
            },
            "type": "array",
            "title": "Audits"
          }
        },
        "type": "object",
        "title": "AuditListResponse"
      },
      "AuditOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "operation": {
            "type": "string",
            "title": "Operation"
          },
          "result": {
            "type": "string",
            "title": "Result"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "operation",
          "result"
        ],
        "title": "AuditOut"
      },
      "DeployBody": {
        "properties": {
          "td": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Td"
          },
          "md": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Md"
          },
          "st_paras": {
            "additionalProperties": true,
            "type": "object",
            "title": "St Paras"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "timeout": {
            "type": "number",
            "title": "Timeout",
            "default": 30.0
          }
        },
        "type": "object",
        "title": "DeployBody",
        "description": "Deployment spec for ``POST /sts/{strategy_id}`` (like deployment.yml)."
      },
      "DeployResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "strategy": {
            "type": "string",
            "title": "Strategy"
          },
          "td": {
            "items": {
              "$ref": "#/components/schemas/TdAttachOut"
            },
            "type": "array",
            "title": "Td"
          },
          "md": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Md"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "live"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "strategy"
        ],
        "title": "DeployResponse"
      },
      "DomainStats": {
        "properties": {
          "domain": {
            "type": "string",
            "title": "Domain"
          },
          "live": {
            "type": "integer",
            "title": "Live",
            "default": 0
          },
          "done": {
            "type": "integer",
            "title": "Done",
            "default": 0
          },
          "healthy": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Healthy"
          }
        },
        "type": "object",
        "required": [
          "domain"
        ],
        "title": "DomainStats"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "SessionListResponse": {
        "properties": {
          "sessions": {
            "items": {
              "$ref": "#/components/schemas/SessionOut"
            },
            "type": "array",
            "title": "Sessions"
          }
        },
        "type": "object",
        "title": "SessionListResponse"
      },
      "SessionOut": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "domain": {
            "type": "string",
            "title": "Domain"
          },
          "created_by": {
            "type": "integer",
            "title": "Created By"
          },
          "created_at": {
            "type": "number",
            "title": "Created At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "api_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Id"
          },
          "sts_session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sts Session Id"
          },
          "strategy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strategy"
          },
          "paused": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paused"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "domain",
          "created_by",
          "created_at",
          "status"
        ],
        "title": "SessionOut"
      },
      "StatsResponse": {
        "properties": {
          "domains": {
            "items": {
              "$ref": "#/components/schemas/DomainStats"
            },
            "type": "array",
            "title": "Domains"
          }
        },
        "type": "object",
        "required": [
          "domains"
        ],
        "title": "StatsResponse"
      },
      "StrategiesResponse": {
        "properties": {
          "strategies": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Strategies"
          }
        },
        "type": "object",
        "required": [
          "strategies"
        ],
        "title": "StrategiesResponse"
      },
      "StsControlResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "paused": {
            "type": "boolean",
            "title": "Paused",
            "default": false
          },
          "strategy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strategy"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "status"
        ],
        "title": "StsControlResponse"
      },
      "TdAttachOut": {
        "properties": {
          "api_id": {
            "type": "integer",
            "title": "Api Id"
          },
          "refcount": {
            "type": "integer",
            "title": "Refcount"
          }
        },
        "type": "object",
        "required": [
          "api_id",
          "refcount"
        ],
        "title": "TdAttachOut"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}